+++ connection.go
@@ -268,9 +268,6 @@ func (c *connection) handleWsMessage(rawMessage []byte, pongTimeoutTimer *time.T
 		return errUnexpectedJSON
 	}
 
-	// 受信したメッセージで message type がパースできたものをログとして保存する
-	c.signalingLog(*message, rawMessage)
-
 	switch message.Type {
 	case "pong":
 		timerStop(pongTimeoutTimer)
@@ -315,6 +312,8 @@ func (c *connection) handleWsMessage(rawMessage []byte, pongTimeoutTimer *time.T
 		c.ayameClient = registerMessage.AyameClient
 		c.environment = registerMessage.Environment
 		c.libwebrtc = registerMessage.Libwebrtc
+
+		c.signalingLog(*message, rawMessage)
 
 		// Webhook 系のエラーログは Caller をつける
 		resp, err := c.authnWebhook()
@@ -388,6 +387,7 @@ func (c *connection) handleWsMessage(rawMessage []byte, pongTimeoutTimer *time.T
 			c.errLog().Msg("RegistrationIncomplete")
 			return errRegistrationIncomplete
 		}
+		c.signalingLog(*message, rawMessage)
 		c.forward(rawMessage)
 	case "message":
 		// 設定が有効でなければ default と同じ処理
@@ -401,6 +401,7 @@ func (c *connection) handleWsMessage(rawMessage []byte, pongTimeoutTimer *time.T
 			c.errLog().Msg("RegistrationIncomplete")
 			return errRegistrationIncomplete
 		}
+		c.signalingLog(*message, rawMessage)
 		c.forward(rawMessage)
 	case "connected":
 		// register が完了していない
@@ -414,6 +415,7 @@ func (c *connection) handleWsMessage(rawMessage []byte, pongTimeoutTimer *time.T
 			c.errLog().Err(err).Send()
 			return err
 		}
+		c.signalingLog(*message, rawMessage)
 	default:
 		c.errLog().Msg("InvalidMessageType")
 		return errInvalidMessageType